requirements.txt
Django~=4.2
djangorestframework~=3.14.0
python-dotenv~=1.0.0
psycopg2-binary